check if the download-lock file exists before doing unlink

+++ cli/prisma2/scripts/download.js
@@ -41,7 +41,9 @@ function createLockFile() {
 function cleanupLockFile() {
   if (createdLockFile) {
     try {
-      fs.unlinkSync(lockFile)
+      if (fs.existsSync(lockFile)) {
+        fs.unlinkSync(lockFile)
+      }
     } catch (e) {
       debug(e)
     }
